New crawler code.

+++ src/main/java/com/weavers/duqhan/domain/UserActivity.java
@@ -5,6 +5,7 @@
  */
 package com.weavers.duqhan.domain;
 
+import java.math.BigInteger;
 import java.util.Date;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -23,7 +24,7 @@ public class UserActivity extends BaseDomain {
 
     private static final long serialVersionUID = 1L;
     @Column(name = "user_id")
-    private Long userId;
+    private BigInteger userId;
     // @Pattern(regexp="[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?", message="Invalid email")//if the field contains email address consider using this annotation to enforce field validation
     @Size(max = 255)
     @Column(name = "email")
@@ -44,11 +45,11 @@ public class UserActivity extends BaseDomain {
     @Column(name = "user_agent")
     private String userAgent;
 
-    public Long getUserId() {
+    public BigInteger getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(BigInteger userId) {
         this.userId = userId;
     }
